ft_printf

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bcili <bcili@student.42.fr>                +#+  +:+       +#+        */
+/*   By: musyilma <musyilma@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 15:11:19 by bcili             #+#    #+#             */
-/*   Updated: 2024/12/02 14:00:15 by bcili            ###   ########.fr       */
+/*   Updated: 2024/12/02 18:18:31 by musyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 int	ft_printf(const char *s, ...);
 int	ft_putchar(int c);
-int	ft_puthex(unsigned long x, char c);
+int	ft_puthex(unsigned int x, char c);
 int	ft_putnbr(int x);
 int	ft_putptr(unsigned long x);
 int	ft_putstr(char *s);
